requirements-dev.txt
-r requirements.txt
pytest>=7.4.0
requests-mock>=1.11.0
